this should maybe work ?

+++ scripts/push-puara-arduino.sh
@@ -1,5 +1,4 @@
 #!/bin/bash -eux
-
 (
   cd puara-arduino || exit 1
 
@@ -8,7 +7,15 @@
 
   git checkout -b continuous
 
+  git add examples/*
   git commit -am "$GITHUB_REF :: $GITHUB_SHA"
 
-  git push --set-upstream origin continuous --force
+  # looks to see if there is something to commit.
+  if [ `git status --porcelain=1 | wc -l` -eq 0 ]; then
+      echo "Nothing to commit."
+      exit 0
+  else
+      git push --set-upstream origin continuous --force
+  fi
+
 )
